make report into service
